requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cors-headers==3.10.0
django-phonenumber-field==6.0.0
djangorestframework==3.12.4
phonenumbers==8.12.38
psycopg2==2.9.2
pytz==2021.3
sqlparse==0.4.2
